requirements.txt
pyserial>=3.5
python-tk>=0.1.0